[IN003] Rename Identifiable interface and abstract class to ApiObject
- Renamed methods.
- Renamed class and interface name.
