+++ hmd-api/Model/IApiObject.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace hmd_api.Model
+{
+    interface IApiObject
+    {
+        public string Id();
+        public abstract string Type();
+    }
+}
